successfully merged create and edit component

+++ src/app/users/index/index.component.ts
@@ -24,6 +24,9 @@ export class IndexComponent implements OnInit{
       .subscribe(data => this.users = data);
   }
   onDelete(id): void {
+    if (!confirm('Are you sure about that?')){
+      return;
+    }
     this.userService.delete(id).subscribe(() => {
       this.ngOnInit();
     });
